Update presets option in .babelrc
This pull request fixes the issue where the Metro bundler in React Native v0.55.3 couldn't find the Babel preset "es2015" and thus fail to bundle JavaScript files. Therefore, the presets option in .babelrc has been updated to "react-native" in order to match the .babelrc that ships with the latest React Native release.